+++ CHANGELOG.md
@@ -4,3 +4,4 @@ Each package has its own changelog.
 - [@ambiki/auto-complete-element](./custom-elements/auto-complete-element/CHANGELOG.md)
 - [@ambiki/clippy-copy-element](./custom-elements/clippy-copy-element/CHANGELOG.md)
 - [@ambiki/combobox](./packages/combobox/CHANGELOG.md)
+- [@ambiki/details-modal-element](./custom-elements/details-modal-element/CHANGELOG.md)